split out datetime functions from `utils` to `_dates`
